+++ src/test/java/LoginUserTest.java
@@ -1,19 +1,15 @@
-import io.qameta.allure.Step;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import json.CreatedUserData;
 import json.LoginData;
 import json.UserData;
+import lib.ApiCore;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-
 import java.util.Random;
-
 import static data.StellarBurgersData.*;
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.notNullValue;
 
 public class LoginUserTest {
     private String testEmail;
@@ -30,25 +26,28 @@ public class LoginUserTest {
     }
     @Test
     public void existedUserLoginTest(){
+        ApiCore apiCore = new ApiCore();
         UserData userData = new UserData(testEmail, testName, testPassword);
-        Response responseCreateUser = postCreateUser(userData);
-        Response responseLoginUser = postLogin(testEmail, testPassword);
-        checkPositiveLoginResponse(responseLoginUser);
+        Response responseCreateUser = apiCore.postCreateUser(userData);
+        Response responseLoginUser = apiCore.postLogin(testEmail, testPassword);
+        apiCore.checkPositiveLoginResponse(responseLoginUser, testEmail, testName);
     }
     @Test
     public void wrongLoginTest(){
+        ApiCore apiCore = new ApiCore();
         UserData userData = new UserData(testEmail, testName, testPassword);
-        Response responseCreateUser = postCreateUser(userData);
-        Response responseNegativeLogin = postLogin( "wrong" + testEmail , testPassword);
-        checkNegativeLoginResponse(responseNegativeLogin);
+        Response responseCreateUser = apiCore.postCreateUser(userData);
+        Response responseNegativeLogin = apiCore.postLogin( "wrong" + testEmail , testPassword);
+        apiCore.checkNegativeLoginResponse(responseNegativeLogin);
 
     }
     @Test
     public void wrongPasswordTest(){
+        ApiCore apiCore = new ApiCore();
         UserData userData = new UserData(testEmail, testName, testPassword);
-        Response responseCreateUser = postCreateUser(userData);
-        Response responseNegativeLogin = postLogin(testEmail,"wrong" + testPassword);
-        checkNegativeLoginResponse(responseNegativeLogin);
+        Response responseCreateUser = apiCore.postCreateUser(userData);
+        Response responseNegativeLogin = apiCore.postLogin(testEmail,"wrong" + testPassword);
+        apiCore.checkNegativeLoginResponse(responseNegativeLogin);
     }
     @After
     public void deleteData(){
@@ -67,39 +66,5 @@ public class LoginUserTest {
                 .header("Authorization",createdUserData.getAccessToken())
                 .when()
                 .delete(DELETE_AUTH_USER);
-    }
-    @Step("Создание пользователя POST /api/auth/register")
-    public Response postCreateUser(UserData userData){
-        return given()
-                .header("Content-Type", "application/json")
-                .and()
-                .body(userData)
-                .when()
-                .post(POST_AUTH_REGISTER);
-    }
-    @Step("Корректная авторизация пользователя POST /api/auth/login")
-    public Response postLogin(String email, String password){
-        LoginData loginData = new LoginData(email, password);
-        return given()
-                .header("Content-Type", "application/json")
-                .and()
-                .body(loginData)
-                .when()
-                .post(POST_AUTH_LOGIN);
-    }
-    @Step("Проверка ответа POST /api/auth/login после корректной авторизации")
-    public void checkPositiveLoginResponse(Response response){
-        response.then().assertThat().body("success", equalTo(true))
-                .and().assertThat().body("accessToken", notNullValue())
-                .and().assertThat().body("refreshToken", notNullValue())
-                .and().assertThat().body("user.email", equalTo(testEmail.toLowerCase()))
-                .and().assertThat().body("user.name", equalTo(testName))
-                .and().statusCode(200);
-    }
-    @Step("Проверка ответа POST /api/auth/login после авторизации с некорректными данными")
-    public void checkNegativeLoginResponse(Response response){
-        response.then().assertThat().body("success", equalTo(false))
-                .and().assertThat().body("message", equalTo(LOGIN_WITH_WRONG_CRED))
-                .and().statusCode(401);
     }
 }
