+++ src/test/java/RegisterUserTest.java
@@ -1,8 +1,5 @@
 import static data.StellarBurgersData.*;
-import static data.StellarBurgersData.USER_ALREADY_EXISTS;
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.notNullValue;
 
 import io.qameta.allure.Step;
 import io.restassured.RestAssured;
@@ -10,6 +7,7 @@ import io.restassured.response.Response;
 import json.CreatedUserData;
 import json.LoginData;
 import json.UserData;
+import lib.ApiCore;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -28,21 +26,21 @@ public class RegisterUserTest {
         testEmail = "testEmail" + random.nextInt(1000000) + "@yandex.ru";
         testName = "testName" + random.nextInt(1000000);
         testPassword = "testPassword" + random.nextInt(1000000);
-
-
     }
     @Test
     public void createUniqueUserTest(){
+        ApiCore apiCore = new ApiCore();
         UserData userData = new UserData( testEmail, testName, testPassword );
-        Response createUserResponse = postCreateUser( userData );
-        checkResponsePostAuthRegister(createUserResponse, testEmail, testName);
+        Response createUserResponse = apiCore.postCreateUser( userData );
+        apiCore.checkResponsePostAuthRegister(createUserResponse, testEmail, testName);
     }
     @Test
     public void createNotUniqueUserTest(){
+        ApiCore apiCore = new ApiCore();
         UserData userData = new UserData( testEmail, testName, testPassword );
-        Response positiveResponse = postCreateUser( userData );
-        Response negativeResponse = postCreateUser( userData );
-        checkNegativeResponsePostAuthRegisterUserExists(negativeResponse);
+        Response positiveResponse = apiCore.postCreateUser( userData );
+        Response negativeResponse = apiCore.postCreateUser( userData );
+        apiCore.checkNegativeResponsePostAuthRegisterUserExists(negativeResponse);
     }
 
     @After
@@ -62,29 +60,5 @@ public class RegisterUserTest {
                 .header("Authorization",createdUserData.getAccessToken())
                 .when()
                 .delete(DELETE_AUTH_USER);
-    }
-    @Step("Создание пользователя POST /api/auth/register")
-    public Response postCreateUser(UserData userData){
-        return given()
-                .header("Content-Type", "application/json")
-                .and()
-                .body(userData)
-                .when()
-                .post(POST_AUTH_REGISTER);
-    }
-    @Step("Проверка ответа POST /api/auth/register. Положительные ответ")
-    public void checkResponsePostAuthRegister(Response response, String testEmail, String testName){
-        response.then().assertThat().body("success", equalTo(true))
-                .and().assertThat().body("accessToken", notNullValue())
-                .and().assertThat().body("user.email", equalTo(testEmail.toLowerCase()))
-                .and().assertThat().body("user.name", equalTo(testName))
-                .and().assertThat().body("refreshToken", notNullValue())
-                .and().statusCode(200);
-    }
-    @Step("Проверка ответа POST /api/auth/register. Пользователь уже создан")
-    public void checkNegativeResponsePostAuthRegisterUserExists(Response response){
-        response.then().assertThat().body("success", equalTo(false))
-                .and().assertThat().body("message", equalTo(USER_ALREADY_EXISTS))
-                .and().statusCode(403);
     }
 }
